+++ docs/changelog.rst
@@ -3,5 +3,6 @@ Changelog
 
 `Next Release`_
 ---------------
+- Add :func:`coercion.stringify`
 
 .. _Next Release: https://github.com/dave-shawley/coercion/compare/0.0.0...master